The global variable space size should be a 16-bit int so we can represent numbers 0-1023

+++ Sources/EV3/EV3Constants.swift
@@ -58,4 +58,4 @@ enum EV3Layer: UInt8 {
 }
 
 let EV3ColorMode = UInt8(2)
-let EV3MaxFileLength = UInt8(64)
+let EV3MaxFileLength = UInt16(64)
